Extract read-only IPC handlers
