+++ packages/app/src/ipc-read-handlers.ts
@@ -1,0 +1,139 @@
+import type { BrowserWindow, IpcMain } from 'electron';
+import type { Logger } from '@invoker/contracts';
+import type { SQLiteAdapter } from '@invoker/data-store';
+import type { AgentRegistry } from '@invoker/execution-engine';
+import type { Orchestrator, TaskDelta, TaskState } from '@invoker/workflow-core';
+import { loadConfig } from './config.js';
+import type { TaskSnapshotCache } from './delta-merge.js';
+
+export interface RegisterReadOnlyIpcHandlersContext {
+  ipcMain: IpcMain;
+  logger: Logger;
+  persistence: SQLiteAdapter;
+  orchestrator: Orchestrator;
+  agentRegistry: AgentRegistry;
+  lastKnownTaskStates: TaskSnapshotCache;
+  getMainWindow: () => BrowserWindow | null;
+  sendTaskDeltaToRenderer: (delta: TaskDelta) => void;
+  setLastKnownWorkflowCount: (count: number) => void;
+  loadTaskByIdFromPersistence: (taskId: string) => TaskState | undefined;
+  resolveAgentSession: (
+    sessionId: string,
+    agentName: string,
+    agentRegistry: AgentRegistry,
+    tasks: TaskState[],
+  ) => Promise<unknown>;
+  timeStartupPhase: <T>(label: string, fn: () => T, fields?: () => Record<string, unknown>) => T;
+  recordStartupDuration: (label: string, startedAtMs: number, fields?: Record<string, unknown>) => void;
+}
+
+export function registerReadOnlyIpcHandlers(context: RegisterReadOnlyIpcHandlersContext): void {
+  const {
+    ipcMain,
+    logger,
+    persistence,
+    orchestrator,
+    agentRegistry,
+    lastKnownTaskStates,
+    getMainWindow,
+    sendTaskDeltaToRenderer,
+    setLastKnownWorkflowCount,
+    loadTaskByIdFromPersistence,
+    resolveAgentSession,
+    timeStartupPhase,
+    recordStartupDuration,
+  } = context;
+
+  ipcMain.handle('invoker:list-workflows', () => persistence.listWorkflows());
+  ipcMain.handle('invoker:get-execution-pools', () => Object.keys(loadConfig().executionPools ?? {}));
+
+  ipcMain.handle('invoker:load-workflow', (_event, workflowId: string) => {
+    logger.info(`load-workflow: "${workflowId}"`, { module: 'ipc' });
+    orchestrator.syncFromDb(workflowId);
+    const tasks = persistence.loadTasks(workflowId);
+    const workflow = persistence.loadWorkflow(workflowId);
+    logger.info(`load-workflow: found ${tasks.length} tasks for "${workflow?.name ?? workflowId}"`, { module: 'ipc' });
+    for (const task of tasks) {
+      lastKnownTaskStates.set(task.id, JSON.stringify(task));
+      const mainWindow = getMainWindow();
+      if (mainWindow && !mainWindow.isDestroyed()) {
+        sendTaskDeltaToRenderer({ type: 'created', task });
+      }
+    }
+    return { workflow, tasks };
+  });
+
+  ipcMain.handle('invoker:get-tasks', (_event, forceRefresh?: boolean) => {
+    const startedAtMs = Date.now();
+    if (forceRefresh) {
+      timeStartupPhase('get-tasks.force-refresh.syncAllFromDb', () => orchestrator.syncAllFromDb(), () => ({
+        taskCount: orchestrator.getAllTasks().length,
+      }));
+    }
+    const tasks = orchestrator.getAllTasks();
+    const workflows = persistence.listWorkflows();
+    if (forceRefresh) {
+      const previousTaskIds = new Set(lastKnownTaskStates.keys());
+      lastKnownTaskStates.clear();
+      for (const task of tasks) {
+        lastKnownTaskStates.set(task.id, JSON.stringify(task));
+        previousTaskIds.delete(task.id);
+        const mainWindow = getMainWindow();
+        if (mainWindow && !mainWindow.isDestroyed()) {
+          sendTaskDeltaToRenderer({ type: 'created', task });
+        }
+      }
+      const mainWindow = getMainWindow();
+      if (mainWindow && !mainWindow.isDestroyed()) {
+        for (const removedTaskId of previousTaskIds) {
+          sendTaskDeltaToRenderer({ type: 'removed', taskId: removedTaskId, previousTaskStateVersion: 0 });
+        }
+        mainWindow.webContents.send('invoker:workflows-changed', workflows);
+      }
+      setLastKnownWorkflowCount(workflows.length);
+    }
+    logger.info(
+      `get-tasks(forceRefresh=${forceRefresh ? 'true' : 'false'}) returning ${tasks.length} tasks, ${workflows.length} workflows`,
+      { module: 'ipc' },
+    );
+    if (forceRefresh) {
+      recordStartupDuration('get-tasks.force-refresh.return', startedAtMs, {
+        taskCount: tasks.length,
+        workflowCount: workflows.length,
+        jsonSizeBytes: Buffer.byteLength(JSON.stringify({ tasks, workflows }), 'utf8'),
+      });
+    }
+    return { tasks, workflows };
+  });
+
+  ipcMain.handle('invoker:get-events', (_event, taskId: string) => persistence.getEvents(taskId));
+  ipcMain.handle('invoker:get-status', () => orchestrator.getWorkflowStatus());
+  ipcMain.handle('invoker:get-task-by-id', (_event, taskId: string) => loadTaskByIdFromPersistence(taskId) ?? null);
+  ipcMain.handle('invoker:get-task-output', (_event, taskId: string) => persistence.getTaskOutput(taskId));
+  ipcMain.handle('invoker:get-output-chunks', (_event, taskId: string) => persistence.getOutputChunks(taskId));
+  ipcMain.handle('invoker:replay-output-from', (_event, taskId: string, fromOffset: number) =>
+    persistence.replayOutputFrom(taskId, fromOffset)
+  );
+  ipcMain.handle('invoker:get-output-tail', (_event, taskId: string) => persistence.getOutputTail(taskId));
+  ipcMain.handle('invoker:get-all-completed-tasks', () => persistence.loadAllCompletedTasks());
+
+  ipcMain.handle('invoker:get-claude-session', async (_event, sessionId: string) => {
+    logger.info(`get-claude-session: "${sessionId}"`, { module: 'ipc' });
+    try {
+      return await resolveAgentSession(sessionId, 'claude', agentRegistry, orchestrator.getAllTasks());
+    } catch (err) {
+      logger.error(`get-claude-session failed: ${err}`, { module: 'ipc' });
+      return null;
+    }
+  });
+
+  ipcMain.handle('invoker:get-agent-session', async (_event, sessionId: string, agentName?: string) => {
+    logger.info(`get-agent-session: "${sessionId}" agent="${agentName ?? 'claude'}"`, { module: 'ipc' });
+    try {
+      return await resolveAgentSession(sessionId, agentName ?? 'claude', agentRegistry, orchestrator.getAllTasks());
+    } catch (err) {
+      logger.error(`get-agent-session failed: ${err}`, { module: 'ipc' });
+      return null;
+    }
+  });
+}
